requirements.txt
Django==1.7
MySQL-python==1.2.5
xlrd==0.9.3
xlwt-future==0.8.0